requirements.txt
Cython==0.26
Pillow==4.2.1
matplotlib==2.0.2
numpy==1.13.1
scipy==0.19.1
scikit-image
scikit_learn==0.19b2
tensorflow==1.3.0rc0
tqdm==4.15.0
transforms3d==0.3.1
PyYAML==3.12
pynrrd
